requirements.txt
youtube_transcript_api
aiofiles
openai-whisper
pyaudio